Notification: refactoring; renamed notificationService to executor, pulled up message building so that execution is independent of API model
